rtx: fix integer overflow in macrocell index decomposition
In computeValueRangesGPU the z-slice stride mcDims.x * mcDims.y was
computed as 32-bit int, overflowing for large uniform grids. Cast
mcDims.x to size_t to match threadID and the numMCs bound above.

+++ devices/rtx/device/spatial_field/space_skipping/UniformGrid.cu
@@ -110,7 +110,7 @@ __global__ void computeValueRangesGPU(box1 *valueRanges,
 
   ivec3 mcID(threadID % mcDims.x,
       threadID / mcDims.x % mcDims.y,
-      threadID / (mcDims.x * mcDims.y));
+      threadID / (size_t(mcDims.x) * mcDims.y));
 
   // Field voxel range covered by this macrocell, with 1-voxel margin
   // to account for trilinear interpolation across macrocell boundaries
